[IMP] Import Reservations
