+++ hotel_channel_connector/components/__init__.py
@@ -6,4 +6,5 @@ from . import backend_adapter
 from . import binder
 from . import importer
 from . import exporter
+from . import deleter
 from . import mapper
