+++ hotel_channel_connector/components/core.py
@@ -9,6 +9,18 @@ class BaseHotelChannelConnectorComponent(AbstractComponent):
     _inherit = 'base.connector'
     _collection = 'channel.backend'
 
+    @api.model
+    def create_issue(self, **kwargs):
+        self.env['hotel.channel.connector.issue'].sudo().create({
+            'backend_id': kwargs.get('backend', False),
+            'section': kwargs.get('section', False),
+            'internal_message': kwargs.get('internal_message', False),
+            'channel_object_id': kwargs.get('channel_object_id', False),
+            'channel_message': kwargs.get('channel_message', False),
+            'date_start': kwargs.get('dfrom', False),
+            'date_end': kwargs.get('dto', False),
+        })
+
 class ChannelConnectorError(Exception):
     def __init__(self, message, data):
         super().__init__(message)
