+++ hotel_channel_connector/models/hotel_reservation/__init__.py
@@ -3,3 +3,4 @@
 
 from . import common
 from . import importer
+from . import exporter
