+++ hotel_channel_connector/models/hotel_room_type/__init__.py
@@ -4,3 +4,4 @@
 from . import common
 from . import importer
 from . import exporter
+from . import deleter
