+++ hotel_channel_connector/models/hotel_room_type_restriction/deleter.py
@@ -1,0 +1,15 @@
+# Copyright 2018 Alexandre Díaz <dev@redneboa.es>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo.addons.component.core import Component
+from odoo import api
+
+class HotelRoomTypeRestrictionDeleter(Component):
+    _name = 'channel.hotel.room.type.restriction.deleter'
+    _inherit = 'hotel.channel.deleter'
+    _apply_on = ['channel.hotel.room.type.restriction']
+    _usage = 'hotel.room.type.restriction.deleter'
+
+    @api.model
+    def delete_rplan(self, binding):
+        return self.backend_adapter.delete_rplan(binding.external_id)
