--- hotel_channel_connector/views/general.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-    <!-- Backend stuff -->
-    <template id="assets_backend" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <script type="text/javascript" src="/hotel_channel_connector/static/src/js/views/list/list_model.js"></script>
-            <script type="text/javascript" src="/hotel_channel_connector/static/src/js/views/list/list_controller.js"></script>
-            <script type="text/javascript" src="/hotel_channel_connector/static/src/js/views/list/list_view.js"></script>
-        </xpath>
-    </template>
-
-</odoo>
